Re-organized UI files, renamed files
