--- WZDCTool/app/src/main/java/com/wzdctool/android/ui/datacollection/DataCollectionViewModel.kt
@@ -1,7 +1,0 @@
-package com.wzdctool.android.ui.datacollection
-
-import androidx.lifecycle.ViewModel
-
-class DataCollectionViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
