+++ WZDCTool/app/src/main/java/com/wzdctool/android/ui/download/DownloadFragmentViewModel.kt
@@ -1,11 +1,10 @@
-package com.wzdctool.android
+package com.wzdctool.android.ui.download
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.wzdctool.android.R
 import com.wzdctool.android.repos.ConfigurationRepository
-import com.wzdctool.android.repos.DataClassesRepository.notificationSubject
-import com.wzdctool.android.repos.DataClassesRepository.toastNotificationSubject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
@@ -14,12 +13,14 @@ class DownloadFragmentViewModel : ViewModel() {
 
     var navigationLiveData = MutableLiveData<Int>()
 
+    // Async. get sorted list of config files
     fun updateCloudConfigFiles() {
         viewModelScope.launch(Dispatchers.IO) {
             ConfigurationRepository.updateCloudConfigList()
         }
     }
 
+    // Async. download specific config files
     fun downloadConfigFiles(configList: List<String>) {
         viewModelScope.launch(Dispatchers.IO) {
             ConfigurationRepository.downloadNewConfigFiles(configList)
