--- WZDCTool/app/src/main/java/com/wzdctool/android/ui/initialization/InitializationViewModel.kt
@@ -1,8 +1,0 @@
-package com.wzdctool.android.ui.initialization
-
-import androidx.lifecycle.ViewModel
-
-class InitializationViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-
-}
