+++ WZDCTool/app/src/main/java/com/wzdctool/android/ui/upload/UploadFragmentViewModel.kt
@@ -1,10 +1,9 @@
-package com.wzdctool.android
+package com.wzdctool.android.ui.upload
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.wzdctool.android.repos.ConfigurationRepository
-import com.wzdctool.android.repos.DataClassesRepository
+import com.wzdctool.android.R
 import com.wzdctool.android.repos.DataFileRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -14,6 +13,7 @@ class UploadFragmentViewModel : ViewModel() {
 
     var navigationLiveData = MutableLiveData<Int>()
 
+    // Asynchronously upload data files, return to main page after upload
     fun uploadDataFiles(fileList: List<String>) {
         viewModelScope.launch(Dispatchers.IO) {
             DataFileRepository.uploadDataFiles(fileList)
