+++ WZDCTool/app/src/main/res/layout/configuration_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ConfigurationFragment"
+    tools:context=".ui.configuration.ConfigurationFragment"
     android:background="@color/background">
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -241,7 +241,7 @@
             android:id="@+id/collectionModeSwitchGuideline"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:checked="true"
+            android:checked="false"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.5"
             app:layout_constraintStart_toStartOf="parent"
