+++ WZDCTool/app/src/main/res/layout/download_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DownloadFragment">
+    tools:context=".ui.download.DownloadFragment">
 
     <TextView
         android:id="@+id/textView"
