+++ WZDCTool/app/src/main/res/layout/editing_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditingFragment">
+    tools:context=".ui.visualization.EditingFragment">
 
     <com.google.android.gms.maps.MapView
         android:id="@+id/mapView"
