+++ WZDCTool/app/src/main/res/layout/editing_selection_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditingSelectionFragment">
+    tools:context=".ui.visualization.EditingSelectionFragment">
 
     <TextView
         android:id="@+id/textView"
