+++ WZDCTool/app/src/main/res/layout/help_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HelpFragment">
+    tools:context=".ui.help.HelpFragment">
 
     <ScrollView
         android:id="@+id/help_sv"
