+++ WZDCTool/app/src/main/res/layout/main_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainFragment"
+    tools:context=".ui.main.MainFragment"
     android:background="@color/background">
 
     <ImageButton
