+++ WZDCTool/app/src/main/res/layout/upload_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UploadFragment">
+    tools:context=".ui.upload.UploadFragment">
 
     <TextView
         android:id="@+id/textView"
